+++ src/SessionMetadata.ts
@@ -1,0 +1,17 @@
+import type {ModeKind} from "./app-server/ModeKind";
+import type {ServiceTier} from "./app-server/ServiceTier";
+import type {Model, Thread} from "./app-server/v2";
+
+export type SessionMetadata = {
+    sessionId: string,
+    currentModelId: string,
+    models: Model[],
+    collaborationMode: ModeKind,
+    modelProvider?: string | null,
+    currentServiceTier?: ServiceTier | null,
+    additionalDirectories: string[],
+}
+
+export type SessionMetadataWithThread = SessionMetadata & {
+    thread: Thread,
+}
